fw-4688, login/logout working via cognito hosted ui
- added env vars to readme
- extracted some hooks and constants
