+++ src/common/constants/misc.js
@@ -14,3 +14,6 @@ export const ABOUT_LINK =
   'https://firstvoices.atlassian.net/wiki/spaces/FIR1/pages/1704813/About+FirstVoices'
 export const SUPPORT_LINK =
   'https://firstvoices.atlassian.net/servicedesk/customer/portals'
+
+// the key for storing a post-login redirect location
+export const ORIGINAL_DESTINATION = 'original_destination'
